refactor: extract start-param routing hook from ClientLayout
